[Add] 배달 등록 점포 분기 처리
